Switch to newer boilerplate format and wording.
The boilerplate for including at the top of files is getting updated to be
project-specific and use a better format for parsing by automated documentation
generators like Doxygen.  Also, the US Postal Service address from which you may
obtain information about the GPL has been replaced by an internet URL because
it's 2017.