ShaderPipeline: refactored from pimpl to NVI
Refactored the ShaderPipeline to use NVI instead of pimpl and added an
OpenGL derivative implementation.  Used that implementation in the
OpenGL Mesh implementation and added to the Font class as a test.

Added unit tests to verify the ShaderPipeline gets called as appropriate
during Mesh draw calls.

Removed the no-longer-used VertexArray class.